+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * Fix compilation on 32-bit architecture
 * Fix Tx.status not being set on error on Commit
+* Fix Listen/Unlisten with special characters
 
 # 2.8.0 (March 18, 2016)